apollo context added

+++ graphql/resolvers/user.resolver.ts
@@ -1,22 +1,19 @@
+import { ApolloContext } from "../../config/apollo.context.ts";
 import * as userService from "../../services/user.service.ts";
-import type {
-  CreateUserInput,
-  UpdateUserInput,
-  User,
-} from "../../types/user.d.ts";
+import type { CreateUserInput, UpdateUserInput } from "../../types/user.d.ts";
 
 export default {
   Query: {
-    users: (_: any, __: any, { currentUser }: { currentUser: User }) => {
+    users: (_: any, __: any, { currentUser }: ApolloContext) => {
+      if (!currentUser) throw new Error("Authentication required");
+
       if (currentUser?.role === 0) return userService.listUsers();
       return [];
     },
-    user: (
-      _: any,
-      { id }: { id: string },
-      { currentUser }: { currentUser: any }
-    ) => {
-      if (currentUser?.role === 0) return userService.getUserById(id);
+    user: (_: any, { id }: { id: string }, { currentUser }: ApolloContext) => {
+      if (!currentUser) throw new Error("Authentication required");
+
+      if (currentUser.role === 0) return userService.getUserById(id);
       return null;
     },
   },
@@ -26,33 +23,41 @@ export default {
     updateUser: (
       _: any,
       { id, input }: { id: string; input: UpdateUserInput },
-      { currentUser }: { currentUser: any }
+      { currentUser }: ApolloContext
     ) => {
-      if (currentUser?.role === 0) return userService.updateUser(id, input);
+      if (!currentUser) throw new Error("Authentication required");
+
+      if (currentUser.role === 0) return userService.updateUser(id, input);
       return null;
     },
     updateUserRole: (
       _: any,
       { id, role }: { id: string; role: number },
-      { currentUser }: { currentUser: any }
+      { currentUser }: ApolloContext
     ) => {
-      if (currentUser?.role === 0) return userService.updateUserRole(id, role);
+      if (!currentUser) throw new Error("Authentication required");
+
+      if (currentUser.role === 0) return userService.updateUserRole(id, role);
       return null;
     },
     updateUserStatus: (
       _: any,
       { id, status }: { id: string; status: number },
-      { currentUser }: { currentUser: any }
+      { currentUser }: ApolloContext
     ) => {
-      if (currentUser?.role === 0)
+      if (!currentUser) throw new Error("Authentication required");
+
+      if (currentUser.role === 0)
         return userService.updateUserStatus(id, status);
       return null;
     },
     deleteUser: (
       _: any,
       { id }: { id: string },
-      { currentUser }: { currentUser: any }
+      { currentUser }: ApolloContext
     ) => {
+      if (!currentUser) throw new Error("Authentication required");
+
       if (currentUser.role === 0) return userService.deleteUser(id);
       return false;
     },
